% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facTrac.R
\name{addAge}
\alias{addAge}
\title{Adds an Age Column to the data set}
\usage{
addAge(x)
}
\arguments{
\item{x}{- a data.frame object to be modified}
}
\value{
Returns the same data frame with an added Age column.
}
\description{
Creates an age column with the formula 2016 - Year.Of.Degree + 22, given
that we assume most B.A's are recieved at age 22.
}
\examples{
x <- system.file("extdata", "FacultyInfo.csv", package = "MyPackage")
addAge(x)
}

