% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createfacTrac.R
\name{createfacTrac}
\alias{createfacTrac}
\title{Creates object of class "fracTrac"}
\usage{
createfacTrac(x)
}
\arguments{
\item{x}{- must be a data.frame}
}
\value{
returns the summary of a facTrac class object through an internal call to summary
}
\description{
Takes a data.frame and converts it to class facTrac and then calls summary on
the newly created object to print the relevant information.
}
\examples{
x <- system.file("extdata", "FacultyInfo.csv", package = "MyPackage")
createfacTrac(x)
}

