% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.facTrac}
\alias{summary.facTrac}
\title{Summary method along with the print method}
\usage{
\method{summary}{facTrac}(object, ...)
}
\arguments{
\item{object}{- an object of class facTrac that will be used to gather the information to calculate
the summary results.}

\item{...}{- this function does not take additional arguments at this time.}
}
\value{
lines of text displaying the call made to create the summarized object as well as then
 printing the resulting summary statistics
}
\description{
Summarizes the given data set and creates a list with the average age of
 the faculty, the oldest, the youngest, and the number of faculty in the given dataset
 and then prints these results in text.
}
\examples{
x <- facSearch("Japanese",2)
object <- createfacTrac(x)
summary(object)
}

